Move SSCK into CL
